Refining taxonomic disparity
summary.by.elements now returns a matrix with the disparity index as
well as a list for all taxa > 2 tips (by default), in which the effect
of each taxon on total tree length is reported. The most egregious
outliers are expected of course to have the largest effects on total
tree length when they are deleted.